+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
-- [#429](https://github.com/poanetwork/nifty-wallet/pull/429) - Allow Ledger signMessage method
-
 ## Current Master
 
+## 5.2.2 Tue Dec 29 2020
+
+- [#437](https://github.com/poanetwork/nifty-wallet/pull/437) - Support getting proxy implementation address by following EIP-1967
+- [#436](https://github.com/poanetwork/nifty-wallet/pull/436) - Allow master copy pattern in importing of proxy contract
+- [#435](https://github.com/poanetwork/nifty-wallet/pull/435) - Allow array input type in contracts interactions
+- [#434](https://github.com/poanetwork/nifty-wallet/pull/434) - Add support of tuple type at interaction with read-only contract methods
+- [#432](https://github.com/poanetwork/nifty-wallet/pull/432) - bump rsk-contract-metadata dependency
+
 ## 5.2.1 Fri Dec 04 2020
+
+- [#429](https://github.com/poanetwork/nifty-wallet/pull/429) - Allow Ledger signMessage method
 
 ## 5.2.0 Fri Nov 13 2020